Separate list utilities from iterable utilities
The `every` and `any` method work for any iterable, not just for lists.
