+++ src/Ast/Sass/Expression/IsCalculationSafeVisitor.php
@@ -13,7 +13,7 @@
 namespace ScssPhp\ScssPhp\Ast\Sass\Expression;
 
 use ScssPhp\ScssPhp\Ast\Sass\Expression;
-use ScssPhp\ScssPhp\Util\ListUtil;
+use ScssPhp\ScssPhp\Util\IterableUtil;
 use ScssPhp\ScssPhp\Value\ListSeparator;
 use ScssPhp\ScssPhp\Visitor\ExpressionVisitor;
 
@@ -56,7 +56,7 @@ final class IsCalculationSafeVisitor implements ExpressionVisitor
 
     public function visitListExpression(ListExpression $node): bool
     {
-        return $node->getSeparator() === ListSeparator::SPACE && !$node->hasBrackets() && \count($node->getContents()) > 1 && ListUtil::every($node->getContents(), fn(Expression $expression) => $expression->accept($this));
+        return $node->getSeparator() === ListSeparator::SPACE && !$node->hasBrackets() && \count($node->getContents()) > 1 && IterableUtil::every($node->getContents(), fn(Expression $expression) => $expression->accept($this));
     }
 
     public function visitMapExpression(MapExpression $node): bool
